Use ticks instead of system time
Building the simulation on system time means that we need to involve a
lot of concurrency in our program, that different runs with non-random
setups can still return different results, and that we wrote an ad-hoc
version of the context package.

Instead of doing this, we can build our simulation around ticks, like in
game programming. We then no longer need any concurrency, which makes
for a simpler program, and non-random runs are reproducible.